winter 2017 midterm progress report updates
This commit updates multiple sections of the design to reflect changes that have occured
during the first half of the develpment phase. The changes center on the simulator sections
because that is where most development has occured. Specifically it discusses UI specifics as
well as the use of the meuus library in favor of epehemeris.js