fix issue where location entry breaks if option not selected
This commit fixes a bug identified by @dakoner where if a user did not
select a location suggestion then the location change did not work.
Now if a user doesn't select a suggestion we try and use that to
get the suggestions that were there and then do a geocode with the
results. It also makes the 'place' var and the geocoder properties
of the view.